% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen_api.R
\name{bacen_api}
\alias{bacen_api}
\title{Conexão com a API do Banco Central}
\usage{
bacen_api(url, httr = TRUE)
}
\arguments{
\item{url}{Um string contendo o URL da API do BACEN para a série desejada.}

\item{httr}{Um valor lógico. Se `TRUE`, a função utiliza o pacote `httr` para a conexão.
Caso contrário, utiliza o pacote `httr2`. O valor padrão é `TRUE`.}
}
\value{
Retorna os dados obtidos da API do BACEN.
}
\description{
Esta função estabelece uma conexão com a API do Banco Central do Brasil (BACEN),
utilizando os pacotes `httr` ou `httr2`, para obter dados em formato JSON e convertê-los
em um formato legível como data frames.
}
\examples{
# Exemplo usando o pacote httr:
url <- "https://api.bcb.gov.br/dados/serie/bcdata.sgs.433/dados?formato=json&dataInicial=01/01/2003&dataFinal=31/12/2023"
dados <- bacen_api(url, httr = TRUE)

# Exemplo usando o pacote httr2:
dados <- bacen_api(url, httr = FALSE)


}
