% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen_form.R
\name{bacen_form}
\alias{bacen_form}
\title{Coleta de Dados da API do Banco Central}
\usage{
bacen_form()
}
\value{
Um data frame contendo as séries temporais coletadas, com colunas para datas
e valores correspondentes às séries especificadas.
}
\description{
Esta função permite que o usuário insira informações para coletar séries temporais
diretamente da API do Banco Central do Brasil (BACEN). Ela orienta o preenchimento de dados
como códigos de séries, nomes das séries e períodos desejados, retornando os dados formatados.
}
\examples{
# Exemplo com múltiplas séries:
# Suponha que o usuário forneça as seguintes informações durante a execução:
# Códigos das séries: c('433', '13005', '24364', '25391')
# Nomes atribuídos: c('ipca_br', 'ipca_for', 'ibc_br', 'ibcr_ce')
# Período: '01/01/2003' a '01/12/2023'

# Após executar a função, os dados podem ser salvos ou utilizados:
dados <- bacen_form()
print(dados)

}
