% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen_search.R
\name{bacen_search}
\alias{bacen_search}
\title{Search for Financial Indicators Series Number}
\usage{
bacen_search(keyword)
}
\arguments{
\item{keyword}{A string containing the search term in Portuguese - BR (e.g., "câmbio", "juros", "Bovespa").}
}
\value{
A data frame with the columns `Code`, `Full_Name`, `Unit`, `Periodicity`, and `Start_Date` corresponding to the search results.

A data frame with the columns `Code`, `Full_Name`, `Unit`, `Periodicity`, and `Start_Date` corresponding to the search results.
}
\description{
Performs a search for financial indicator series based on a keyword.
}
\examples{
# Search for indices related to Fortaleza
bacen_search("fortaleza")

# Search for indices related to Fortaleza
bacen_search("fortaleza")

}
