% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen_url.R
\name{bacen_url}
\alias{bacen_url}
\title{Criação de URLs para acesso à API do Banco Central.}
\usage{
bacen_url(serie, data_inicio, data_termino)
}
\arguments{
\item{serie}{Código(s) da(s) série(s) desejada(s). Deve ser um número ou vetor de números.}

\item{data_inicio}{Data inicial do período de interesse, no formato "dd/mm/aaaa".}

\item{data_termino}{Data final do período de interesse, no formato "dd/mm/aaaa".}
}
\value{
Retorna um vetor contendo os URLs criados para cada série fornecida.
}
\description{
Esta função gera URLs que permitem acessar dados de séries temporais específicas
fornecidas pela API do Banco Central do Brasil (BACEN).
}
\examples{
# Gerar URL para a série 433 (IPCA) no período de 01/01/2003 a 31/12/2023.
bacen_url(433, "01/01/2003", "31/12/2023")

}
