% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen_form.R
\name{bacen_form}
\alias{bacen_form}
\title{Data Collection from the Central Bank API}
\usage{
bacen_form()
}
\value{
A data frame containing the collected time series, with columns for dates
and values corresponding to the specified series.
}
\description{
This function allows the user to input information to collect time series
directly from the Central Bank of Brazil (BACEN) API. It guides the user
in providing data such as series codes, series names, and desired periods,
returning the data in a formatted structure.
}
\examples{
# Example with multiple series:
# Assume the user provides the following information during execution:
# Series codes: c('433', '13005', '24364', '25391')
# Assigned names: c('ipca_br', 'ipca_for', 'ibc_br', 'ibcr_ce')
# Period: '01/01/2003' a '01/12/2023', in the format "dd/mm/yyyy"

# After running the function, the data can be saved or used:
data  <- bacen_form()
print(data)

}
