% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen_url.R
\name{bacen_url}
\alias{bacen_url}
\title{Creation of URLs to Access the Central Bank API}
\usage{
bacen_url(series, start_date, end_date)
}
\arguments{
\item{series}{Desired series code(s). Must be a number or a vector of numbers.}

\item{start_date}{Start date of the period of interest, in the format "dd/mm/yyyy".}

\item{end_date}{End date of the period of interest, in the format "dd/mm/yyyy".}
}
\value{
Returns a vector containing the generated URLs for each provided series.
}
\description{
This function generates URLs to access specific time series
data provided by the Central Bank of Brazil API.
}
\examples{
# Generate a URL for series 433 (IPCA) from 01/01/2003 to 31/12/2023.
bacen_url(433, "01/01/2020", "31/12/2023") # in the format "dd/mm/yyyy"

}
